Add default option to only run configuration updates for ocata/pike
This is an attempt to skip costly installation steps in incremental
upgrades from Newton to Queens.  Hopefully it works.

+++ scripts/ubuntu16-newton-to-ocata.sh
@@ -18,11 +18,16 @@ set -evu
 
 export RPC_BRANCH=${RPC_BRANCH:-'ocata'}
 export OSA_SHA="stable/ocata"
+export SKIP_INSTALL=${SKIP_INSTALL:-"yes"}
 
 # configure apt sources and update all Xenial Packages before jump to Ocata
 pushd /opt/rpc-upgrades/playbooks
   openstack-ansible prepare-ocata-upgrade.yml
 popd
+
+if [[ "$SKIP_INSTALL" == "yes" ]]; then
+  exit 0
+fi
 
 pushd /opt/rpc-openstack
   git clean -df
